Renamed what needed to be renamed in Rider
