+++ Assets/Scripts/npcscript.cs
@@ -6,12 +6,12 @@ using UnityEngine.SceneManagement;
 
 public class npcscript : MonoBehaviour
 {
-    private NPCScreenInteractor interactor;
+    private DialogueManager interactor;
 
     private void Start()
     {
-        interactor = GameObject.Find("Canvases").GetComponentInChildren<NPCScreenInteractor>(true);
-        // interactor = GameObject.Find("Interactor").GetComponent<NPCScreenInteractor>();
+        interactor = GameObject.Find("Canvases").GetComponentInChildren<DialogueManager>(true);
+        // interactor = GameObject.Find("Interactor").GetComponent<DialogueManager>();
     }
 
     /// Update brings back the previous scene when player presses the escape key
